# api/models/error_data.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// ErrorData Все плохо
// swagger:model Error_data
type ErrorData struct {
	BaseData

	// data
	// Required: true
	Data interface{} `json:"data"`
}

// UnmarshalJSON unmarshals this object from a JSON structure
func (m *ErrorData) UnmarshalJSON(raw []byte) error {
	// AO0
	var aO0 BaseData
	if err := swag.ReadJSON(raw, &aO0); err != nil {
		return err
	}
	m.BaseData = aO0

	// AO1
	var dataAO1 struct {
		Data interface{} `json:"data"`
	}
	if err := swag.ReadJSON(raw, &dataAO1); err != nil {
		return err
	}

	m.Data = dataAO1.Data

	return nil
}

// MarshalJSON marshals this object to a JSON structure
func (m ErrorData) MarshalJSON() ([]byte, error) {
	_parts := make([][]byte, 0, 2)

	aO0, err := swag.WriteJSON(m.BaseData)
	if err != nil {
		return nil, err
	}
	_parts = append(_parts, aO0)

	var dataAO1 struct {
		Data interface{} `json:"data"`
	}

	dataAO1.Data = m.Data

	jsonDataAO1, errAO1 := swag.WriteJSON(dataAO1)
	if errAO1 != nil {
		return nil, errAO1
	}
	_parts = append(_parts, jsonDataAO1)

	return swag.ConcatJSON(_parts...), nil
}

// Validate validates this error data
func (m *ErrorData) Validate(formats strfmt.Registry) error {
	var res []error

	// validation for a type composition with BaseData
	if err := m.BaseData.Validate(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateData(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *ErrorData) validateData(formats strfmt.Registry) error {

	if err := validate.Required("data", "body", m.Data); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *ErrorData) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ErrorData) UnmarshalBinary(b []byte) error {
	var res ErrorData
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}